Added product routes
Product routes are added in the routes folder
In server.js we used app.use(productRoutes) to call the routes
