+++ app/models/user.js
@@ -2,10 +2,6 @@ const mongoose = require('mongoose')
 const orderSchema = require('./order')
 
 const userSchema = new mongoose.Schema({
-  name: {
-    type: String,
-    required: false
-  },
   email: {
     type: String,
     required: true,
