+++ app/routes/product_routes.js
@@ -1,0 +1,69 @@
+// Express docs: http://expressjs.com/en/api.html
+const express = require('express')
+// Passport docs: http://www.passportjs.org/docs/
+const passport = require('passport')
+
+// pull in Mongoose model for products
+const Product = require('../models/product')
+
+// this is a collection of methods that help us detect situations when we need
+// to throw a custom error
+const customErrors = require('../../lib/custom_errors')
+
+// we'll use this function to send 404 when non-existant document is requested
+const handle404 = customErrors.handle404
+
+// passing this as a second argument to `router.<verb>` will make it
+// so that a token MUST be passed for that route to be available
+// it will also set `req.user`
+const requireToken = passport.authenticate('bearer', { session: false })
+
+// instantiate a router (mini app that only handles routes)
+const router = express.Router()
+
+// INDEX
+// GET /products
+router.get('/products', requireToken, (req, res, next) => {
+  Product.find()
+    .then(products => {
+      // `products` will be an array of Mongoose documents
+      // we want to convert each one to a POJO, so we use `.map` to
+      // apply `.toObject` to each one
+      return products.map(product => product.toObject())
+    })
+    // respond with status 200 and JSON of the products
+    .then(products => res.status(200).json({ products: products }))
+    // if an error occurs, pass it to the handler
+    .catch(next)
+})
+
+// CREATE
+// POST /products
+router.post('/products', requireToken, (req, res, next) => {
+  Product.create(req.body.product)
+    // respond to succesful `create` with status 201 and JSON of new "product"
+    .then(product => {
+      res.status(201).json({ product: product.toObject() })
+    })
+    // if an error occurs, pass it off to our error handler
+    // the error handler needs the error message and the `res` object so that it
+    // can send an error message back to the client
+    .catch(next)
+})
+
+// DESTROY
+// DELETE /products/5a7db6c74d55bc51bdf39793
+router.delete('/products/:id', requireToken, (req, res, next) => {
+  Product.findById(req.params.id)
+    .then(handle404)
+    .then(product => {
+      // delete the product if it exists
+      product.deleteOne()
+    })
+    // send back 204 and no content if the deletion succeeded
+    .then(() => res.sendStatus(204))
+    // if an error occurs, pass it to the handler
+    .catch(next)
+})
+
+module.exports = router
